requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
sqlalchemy>=2.0.25
pydantic>=2.5.3
python-multipart>=0.0.6
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-docx>=0.8.11
playwright>=1.40.0
requests>=2.31.0
schedule>=1.2.0
httpx>=0.26.0
apscheduler>=3.10.0
imapclient>=3.0.1
secure-smtplib>=0.1.1 